+++ src/tk/wurst_client/features/mods/ClickAuraMod.java
@@ -17,6 +17,7 @@ import tk.wurst_client.settings.SliderSetting;
 import tk.wurst_client.settings.SliderSetting.ValueDisplay;
 import tk.wurst_client.utils.EntityUtils;
 import tk.wurst_client.utils.EntityUtils.TargetSettings;
+import tk.wurst_client.utils.RotationUtils;
 
 @Mod.Info(
 	description = "Automatically attacks the closest valid entity whenever you click.\n"
@@ -130,7 +131,7 @@ public class ClickAuraMod extends Mod implements UpdateListener
 	public void onDisable()
 	{
 		wurst.events.remove(UpdateListener.class, this);
-		EntityUtils.lookChanged = false;
+		RotationUtils.lookChanged = false;
 	}
 	
 	@Override
@@ -152,8 +153,8 @@ public class ClickAuraMod extends Mod implements UpdateListener
 		Entity entity = EntityUtils.getClosestEntity(targetSettings);
 		
 		// head rotation
-		EntityUtils.lookChanged = entity != null;
-		if(!EntityUtils.lookChanged)
+		RotationUtils.lookChanged = entity != null;
+		if(!RotationUtils.lookChanged)
 			return;
 		
 		// AutoSword
@@ -163,7 +164,7 @@ public class ClickAuraMod extends Mod implements UpdateListener
 		wurst.mods.criticalsMod.doCritical();
 		
 		// face entity
-		if(!EntityUtils.faceEntityPacket(entity))
+		if(!RotationUtils.faceEntityPacket(entity))
 			return;
 		
 		// attack entity
