+++ src/tk/wurst_client/features/mods/KillauraMod.java
@@ -19,6 +19,7 @@ import tk.wurst_client.settings.SliderSetting;
 import tk.wurst_client.settings.SliderSetting.ValueDisplay;
 import tk.wurst_client.utils.EntityUtils;
 import tk.wurst_client.utils.EntityUtils.TargetSettings;
+import tk.wurst_client.utils.RotationUtils;
 
 @Mod.Info(
 	description = "Automatically attacks entities around you.\n"
@@ -225,7 +226,7 @@ public class KillauraMod extends Mod implements UpdateListener
 	public void onDisable()
 	{
 		wurst.events.remove(UpdateListener.class, this);
-		EntityUtils.lookChanged = false;
+		RotationUtils.lookChanged = false;
 	}
 	
 	@Override
@@ -243,8 +244,8 @@ public class KillauraMod extends Mod implements UpdateListener
 		Entity entity = EntityUtils.getClosestEntity(targetSettings);
 		
 		// head rotation
-		EntityUtils.lookChanged = entity != null;
-		if(!EntityUtils.lookChanged)
+		RotationUtils.lookChanged = entity != null;
+		if(!RotationUtils.lookChanged)
 			return;
 		
 		// AutoSword
@@ -254,7 +255,7 @@ public class KillauraMod extends Mod implements UpdateListener
 		wurst.mods.criticalsMod.doCritical();
 		
 		// face entity
-		if(!EntityUtils.faceEntityPacket(entity))
+		if(!RotationUtils.faceEntityPacket(entity))
 			return;
 		
 		// attack entity
